requirements.txt
requests
Pillow
opencv-python-headless